Remove Session.from_csv and related functionality
Removes `Session.from_csv` and all of the other CSV-related methods in Sessions, and replaces their uses in the tutorials and examples.

Also adds a job to run the examples as part of the tests, as they were never getting tested before, and removes a couple of examples that are no longer needed.

tumult#1670