change colab link to sandbox mode
Change the colab link to use "sandbox mode".

For users without permission to edit the colab, the link will look the same.

For users who are logged into a Google account with the ability to edit the colab, there will be a big banner at the top warning you that changes will not be saved.

This should prevent inadvertent edits to the colab.