Enable aggregations on IDs tables with MaxRowsPerID constraint
Allows performing aggregations on tables with the AddRowsWithID protected change
when they have a MaxRowsPerID constraint on them by truncating the table before
aggregating. In the process, adds support for tracking ID columns to Schema and
Catalog, and logic for doing so to the OutputSchemaVisitor. Also improves some
error messages produced by the OutputSchemaVisitor.

tumult#2252