Have `check_installation` error message be the last output when it fails
Changes how `check_installation` handles exceptions so that the diagnostic
message we show the user is the last thing that gets printed out, where
previously it was shown before the traceback that caused the error. Also gives a
slightly more helpful error message for the case where Java isn't installed, and
removes a couple of references to a section of the installation guide that
doesn't exist anymore.

tumult#2362