Unregister cleanup hook when Spark init fails in install check
If `SparkSession.builder.getOrCreate()` fails, the Core `atexit` hook for
cleaning up the materialization database produces a warning (prior to core!185,
an exception). In most cases this is fine, but it makes the installation check
output a bit harder to understand, so this MR unregisters that hook specifically
when getting a Spark session fails in that case. The installation check exits
when that happens, so no queries are run and thus no queries are materialized
that will need to be cleaned up.

tumult#2362