Always include NaN bin in BinningSpec.bins(), if one was specified
Changes the behavior of `BinningSpec.bins()` to include the NaN bin (if any) in
all cases, not just when the bin edges are floats. Also causes the autogenerated 
bins for string bin edges to quote the edges, and reorders the `BinningSpec`
tests slightly to keep similar tests together.

tumult#2699

See merge request tumult-labs/analytics!230