Increase package job timeout
The nightly pipeline has been occasionally failing due to a timeout during the
`package` job, it appears because the executor instance is taking several
minutes to spawn. Exactly why that is remains unclear, but bumping the timeout
should prevent the failure.

Also fixes a minor typo in the docs.

See merge request tumult-labs/analytics!256