Fixes Documentation for Functions that have Automatic Column Identification
This MR adjusts documentation for functions that have Automatic Column Identification. This is specifically an issue when the `protected_change` is `AddRowsWithID` because the `*privacy_ID*` column will have specific rules. For instance, it is required in joins to use the `*privacy_ID*` column, but it cannot be used as a column in the `get_groups` function. Adjusting the documentation in this issue will make the API reference more usable for users.

Issue: https://gitlab.com/tumult-labs/tumult/-/issues/3013

See merge request tumult-labs/analytics-upstream!17

Upstream-hash: f97037892fc1e8dc6c5a4e797db8431086727c08