Refactor MeasurementVisitor
The primary goal was to update the MeasurementVisitor so that all groupby queries used the same format where

1. they had build methods that could be overwritten, and
2. they used a measurement to get a keyset (potentially 0-dp, inf-dp, or approx-dp w/ finite budget)

This new adaptive format also broke the session.\_noise_info infrastructure, but that was actually pretty straightforward to fix by having the MeasurementVisitor & QueryExprCompiler return the noise information in addition to the built measurement.

There were also a ton of tests which were super-overfit to the exact implementation details which were horribly broken with the refactor. I deleted most of them, but replaced a lot of the important ones with new tests.

There are still a massive number of tests which survived the refactor, which give me a lot of confidence that this isn't totally breaking everything :grinning:

See merge request tumult-labs/analytics-upstream!35

Upstream-hash: 11f1624882d78aede52c7bb7509acde1a8e25503