Ignore the nightly_handler job status in the handler script
The pipeline was showing as failed in Slack even though everything passed
because the nightly handler job is running when it runs (yes, that's as obvious
as it sounds), and as a result the handler was seeing it as "not succeeded" and
marking the pipeline as failed.

Also prints out the status of the failed jobs in the handler log; you can get
this information from GitLab fairly easily, but this puts it all in one place if
you open up the job log for the handler job.

Upstream-hash: d309ffa9e2265c0c3c7af85cd12d87d067ccd655