Fix docs version banner after pydata-sphinx-theme upgrade
Changes in pydata-sphinx-theme make it so that jquery is no longer available in
our custom JS snippet for adding banners to the docs, which broke it. This
removes all use of jquery, and also fixes a minor bug with how errors are
handled if the banner configuration fails to load.

Upstream-hash: 4fd514e5c57f6c99b4779353638cdb8b8ade7d55