Fix docs footer layout
The way the footer was structured previously was to have three lines of
information on the left, and the right blank. The recent pydata-sphinx-theme
update changed it to have three lines on the left and one on the right, with
more space between the lines, making the whole thing look unbalanced. This
rearranges it so that information about us (copyright, docs build info) is on
the left, and information about Sphinx (version number, theme) is on the right.

Upstream-hash: 3f370b453e0eaa9aa4da328351c2463aae823fc5