Fix version switcher version number display/matching
Updating pydata-sphinx-theme changes the way the version switcher displays the
current version number, which caused it to not display correctly in some
cases. This changes the version switcher configuration to match on the minor
version number, and also writes out the version and release numbers to files in
the Sphinx output directory so we can use them for automation in the
docs.tmlt.dev repo.

Upstream-hash: 3687432f90e385b00032e863c676b78b1bb79e73