Remove Sphinx .doctrees before publishing docs artifact in CI
The .doctrees directory is used by Sphinx as a build cache, but it's not needed
to actually view the docs, and it's quite large. This change was motivated by
hitting the site size limit on GitLab pages, largely due to having this giant
build cache sitting around for every single version of the docs.

Upstream-hash: 456a8fbcdfbcbe151263f36cc4741e8afb340a91