Return dict, not Schema, from Session.get_schema and KeySet.schema
This changes Session.get_schema and KeySet.schema to return a dictionary of
column names to ColumnDescriptors, rather than a Schema object. The Schema type
is technically private and no longer appears in the docs at all, so returning it
is confusing to users (how do they know what to do with it?) and inconsistent
with the places we ask the user to provide a schema. Additionally, the Schema
being returned by Session.get_schema was subtly wrong, as it did not include the
ID column or grouping column of the table even though that information is
included in the Schema type and the ID space *was* included.

Upstream-hash: 89f2df4b191cfe2ff0254454dfcdf3f21f0fea50