Add storage interface for use with workspaces
Introduces a new StorageInterface ABC and an implementation for it on the local
filesystem, allowing dataframes and arbitrary data to be saved and loaded later.
It is not used for anything yet, but will be a piece of the new workspaces
features.

Upstream-hash: 097316422611b3d2e732a710cf831b3925d69167