Drop support for Python 3.8, PySpark <3.3.1
We announced in 0.17.0 that we were dropping support for Python 3.8 and PySpark
versions below 3.3.1, neither of which are officially supported anymore.

Also removes the no-longer-needed pyarrow_hotfix import, as we no longer depend
on the affected versions of PyArrow, and pulls in some upgrades to the linters
and other dev dependencies. In particular, it increases the minimum supported
version of seaborn, where our previously-used version was having trouble with
backwards-incompatible changes in the most recent version of matplotlib.

Upstream-hash: c0f5228983ac2e1932e51dafde306e7a9bbbc012