Implement KeySet v2 __mul__
Implements `KeySet * KeySet` for the v2 implementation, adding a new `CrossJoin`
`KeySetOp` representing this operation. Also adds a new method, `size`, to
`KeySetOp` that is used for optimizing Spark partitioning in certain cases. Note
that its behavior is different from the existing `KeySet.size` in that it only
computes the size if there is a fast way of doing so, and returns None
otherwise. It is not exposed in the `KeySet` API.

Implementation-wise, this works similarly to the existing KeySet multiplication
code except that:
- A CrossJoin is always between two KeySets in the new implementation, and
  a chain of them does not automatically get flattened into a single operation.
- The partition control logic is slightly different, though in most cases it
  behaves equivalently. One important exception here is that it will
  automatically repartition single-partition factors in some cases to ensure
  that crossjoining a bunch of small factors with one partition each doesn't end
  up with one giant dataframe that still has a single partition.

The thresholds for coalescing/repartitioning before crossjoining weren't chosen
in a particularly scientific way, and it's possible they could be improved. They
seem to provide a reasonable level of performance and scalability in the tests I
ran, though.

Upstream-hash: d29ede93e262c075185c20bf76e8bce98fdbe8cc