Add a defined column order to v2 KeySets
Previously, the column order for a KeySet was treated as an implementation
detail. This made some types of tests tricky to write, and could be confusing to
users in some cases (e.g. if they passed in a hierarchical KeySet and the
columns in the hierarchy were rearranged to not be in order). With this change,
the behavior is better-defined: the KeySet keeps track of a user-facing column
order that has some defined rules, and KeySetOps explicitly disregard column
order.

Specifically, the rules for the currently-existing operations are as follows:
- For from_tuples and detect, the column order is the same as the one the user
  passed in.
- For __mul__, the order is the left factor's columns (in order) followed by the
  right factor's.
- As a result of these, from_dict also produces a KeySet whose column order is
  the order the user specified the column domains in.
As more operations are added, additional rules like this will be needed, though
all of the ones we anticipate adding have fairly obvious behaviors in this
regard.

Upstream-hash: 3bdf4610cc1e03121e163b0e36981ccc6f0eace1