Shrink v2 KeySet cross-join test
The 3^20 cross-join/from_dict test was occasionally failing with
OutOfMemoryError (e.g. [1]), likely because the output is too large to fit in
memory. It now only tests 18 factors instead of 20, which should shrink the
output to be more in line with the other parameter combinations for that test.

[1] https://gitlab.com/tumult-labs/analytics-ee/-/jobs/8595677339

Upstream-hash: 877ab85de46be564ffedaf420e3e109f55655d6f