Implement size method on v2 KeySets
Adds KeySet.size() to v2 KeySets, matching the API of the existing KeySet
implementation. In the process, changes the size() method on KeySetOps to
include an argument, fast, which controls whether it will guarantee an exact
answer at the cost of potentially being slow. For most optimization purposes,
fast is True, but for computing the KeySet size to give to the user it is False.

Upstream-hash: 43f685577e5a103536822ddd8fa0122d28b288ef