Standardize token variable names in CI
At some point we dropped the `ANALYTICS_` prefix to a bunch of CI variable
names, but not all of them, and the intermediate state resulted in a bunch of
duplicate variables. This converts all of the variables to use the newer one so
we don't need to worry about that anymore.

Upstream-hash: a82e2386ae49830f40fe6f9541f45a0d978312a8